✨(front) display a suggestion for the full text search through courses
The course search bar in the courses page only offers suggestion for
courses of resource-based filters. We also intend to support full-text
search through both course title, metadata & syllabus.

Add a special-case "default" suggestion that simply displays the current
content of the fields along with a cue "Search for foo".