✨(front) handle full text queries in courses search
Handle the full text search suggestion in the main courses search. Just add
a case to the `onSuggestionSelected` callback from `react-autosuggest`.

When this suggestion is selected, we should simply trigger a new
getResourceList action with updated params. This is done throug the
container component which passes a handler for this to <SearchSuggestField />.