♻️(search) change the shape of indexers to be more generic
Indexers are supposed to all have the same shape. This is mostly
because they are one of our APIs as they allow developers to change
behaviors in richie.

It can also be useful to enable code reuse inside the project itself.
The format_es_{foo}_for_api methods on the indexers all contained the
name of the related model. This was unncessary and prevented us from
just loading an indexer and using the method.

Change the name so that method can be used without knowing the
indexer's name. Change the list sorting script & related method
in the same way.