✨(courses) add "page_placeholder" template tag
We want to reuse the same template to display the course content on
the course detail page and on the course run detail page. For this
we need a template tag that works like the "placeholder" template
tag but allowing us to specify the page from which the placeholder
should be taken. When the specified page happens to be the current
page, the "page_placeholder" shoud behave like the "placeholder"
template tag. When the specified page is different from the current
page, the "page_placeholder" should behave like the "show_placeholder"
template tag and just display its content as readonly.

We also added the possibility to assign the output of the
"page_placeholder" template tag to a context variable, which was not
possible on the "placeholder" tag. This will be useful to avoid
displaying empty blocks when the output of the tag is null thanks to
the following pattern:
{% page_placeholder "slot" page as var %}
{% if var %}
<h1>{{ var }}</h1>
{% endif %}