♻️(courses) stop including language marker in strings for testing
We automated the inclusion of a language marker when generating
random text for pages in tests. This was awkward and made it difficult
to simply test the presence of a precise string. If what we want to
test is that languages are working as expected, then setting a different
string by language should be handled in the test, not automatically.