♻️(courses) make the course run a subpage of its related course
This change is required so that when a course is being modified,
the course runs associated to the course can still display the
version at the time of the snapshot. A snapshot is a copy of the
course that is placed as a child of the current course page. Course
run pages are children of the course page in a specific version
(either the current course page or one of the snapshots). We will
use the parent page/child page relationship baked in the CMS instead
of a course foreign key so that the current course can easily display
all courseruns by fetching all its descendants.