✨(courses) add a wizard to easily create a new course run
The dates of the course run are not created via the wizard
because the Django admin datetime widget does not work out of
the box. We would need to spend more time to make it work but
for the moment it is ok to set the dates directly on the course
run page because we have activated easy front end editing.