♻️(search) add a _meta attr on ViewSets to store the relevant indexers
In order to promote code reuse in ViewSets through mixins, we need
our ViewSets to provide a way for those mixins to access indexers
(as they hold all resource-specific information and are required to
do basically anything with our ES indices).

This is obviously not possible if they're just a static variable in
the Viewset file. Simply setting them on ViewSets (as eg.
`CoursesViewSet.indexer`) as we previously did was breaking as it
conflicted with some DRF expectations.

We decided to add a `_meta` attribute on ViewSets using a specific
class to mimick Django's `class Meta` and house all relevant
information for the specific ViewSet, starting with the indexer.