🔍(courses) ensure course runs and course snapshots are not indexed
A course is run several times and each run is a separate page but we only
want the main page of the course to be indexed so that a course run that
is past does not end-up in the first page of search results. Course run may
relate to older versions of the course that are materialized by a snapshot
of a course page. For the same reason, we should make sure that these
snapshots are also not going to be indexed by search engine.