♻️(front) use an Enum for model names
We used simple string literals for model names until now. This resulted
in a lot of "magic strings" floating around in the code, which can be
confusing.
It also forced us to do some manual typecasting in certain situations.

In our opinion, enums are the way to handle these kinds of fixed values
that exist both as types & runtime values. We get a clear source of
truth for all use cases related to the model name and the best possible
type checking for these values.