📌(dependencies) update yarn.lock to deduplicate @type/react versions
There is currently an issue in yarn where it fails to deduplicate
dependencies versions on `yarn install`. See this issue:
https://github.com/yarnpkg/yarn/issues/6070.

This, coupled with a change in the definition of Component in
@types/react, (https://github.com/DefinitelyTyped/DefinitelyTyped/commit/9039892f8ea3823b0a37d2b9764ba413a9eb6aa9)
breaks us as it forces us to use incompatible versions of @types/react.

We can easily solve this by unlocking all our dependencies (by
deleting node_modules & yarn.lock) and regenerating the lockfile
from scratch. This is stable as it does not create a different
lockfile than what yarn does on its own.

gitlint-ignore: body-max-line-length