🐛(courses) exclude snapshots from courses on organization and subject
A filter query was missing to exclude snapshots when retrieving the list
of courses from an organization (resp. subject). Note that adding this
filter created a duplicate result issue because the query is spanning
many-to-many relationships. We had to add a "distinct" clause to solve
it and we made sure that the tests were securing this issue (see creation
of a root page/course page/snapshot tree structure in the tests).