✨(courses) add admin method to create course snapshots
This method automatically makes a copy of the course with its extensions,
and places it below the original course. It creates new title objects for
each language and sets them to the current timestamp so that each snapshot
has a unique name.

It then moves all course runs of the course below its
snapshot so we keep track of the version in which the course was when these
course runs were created. Finally, it makes sure to respect the publication
status of the course in each language on the snapshot.

A snapshot can only be triggerd by a user with update and add access rights
on the original course.