💄(courses) display course title on snapshots and course runs
Courses might be main courses or snapshot. If they are snapshots, the
meaningful title to display is the title of the course (the title of
the snapshot is just a timestamp for admin) and it should be readonly.

Course runs may be below a course or below a snapshot. If they are
below snapshots, the meaningful title to display is the title of the
course and it should be readonly.