👌(courses) remove snapshots from course plugin choices
Code review revealed that snapshots were showing up among the choices.
Added conditions to filter them out. Added test scenario to make sure.
Unintended consequence is that only courses that have a parent can be
added by course plugin since we need to check for the absence of a
course page as parent. The absence of a parent filters it out.