✨(courses) add a languages field on the course run
We want to be able to indicate, on the course syllabus, the
languages in which the course content is available. I added
a required Array field on the course run object and displayed
the field on the course and on the course run detail views
with frontend editing enabled.

Note that the use of an ArrayField forces the use of Postgresql.