✨(courses) allow all languages not just CMS languages on the course run
The LANGUAGE setting is used to designate languages that are activated on
the CMS. We want to allow selecting any languages on the course run, even
if it is not an active language on the CMS. I had to rename some factory
fields that are related to pages because one of them was named "languages".
I renamed them to "page_*" which makes things cleare and avoids later
conflicts.