⬆️(dependencies) update yarn.lock to drop flatmap-stream
`flatmap-stream` was recently revealed to be a vehicle for a
malicious code injection. It was subsequently unpublished.

As it was part of our lockfile, this broke our builds. We can
regenerate the lockfile to drop it (through `event-stream`, which
included it in the first place) and get our builds going again.