📌(dependencies) use a resolution to pin `@types/react`
We were having trouble with incompatible versions of `@types/react`
getting pulled in through our transitive dependencies.

We decided to pin the version through a resolution to prevent any
further inconvenience from this issue. We'll be able to remove it
once all dependencies have updated their own version of the react
types.