🐛(docker) fix default command
When we've refactored the project tree, we forgot to fix the path to our
wsgi configuration to run gunicorn.

We choose to change to the sandbox directory before running gunicorn to
prevent PYTHONPATH-related issues (e.g. to import settings). Another
strategy would have been to change the WORKDIR, but it also means that
this would break our development images building, and it would be
disturbing to default to the sandbox directory instead of the project's
root.