♻️(courses) create Elasticsearch index of courses from model
Up to now, the Elasticsearch index for courses was generated from
calls to the fun-mooc.fr API. It was only used for development
purposes. We now want to index the real data in our courses model.
This is a first working implementation with minimal content:
title, description, dates, organizations, subjects and cover image.