🐛(makefile) force base image rebuild
When forcing the app image rebuild using the 'make rebuild' command,
only the "app" image was rebuilt and not the "base" image, leading to
inconsistencies when adding new dependencies for example. By forcing to
rebuild the base image and then the app image, we ensure that the app
image is build upon the base image.