✨(search) add links to course runs in Search
We could not link to the courses / course runs from the course
glimpses in seach before actual courses from the models were
indexed in elasticsearch. Ditto for the courses in suggestions.

Now that they are, we can trivially add those links by using the
new `absolute_url` field on Courses.