🐛(search) add full text search to history
We were adding filters (such as facets or related resources) to the
URL and history, but not full text search queries. They were only
added to the URL when eg. a filter forced an update.

We simply needed to add the relevant action to the handler that
we use from <SearchSuggestField /> in <SearchSuggestFieldContainer />.

As we did, we also noticed the UI was no picking up the "query" param
from the URL. We then added a call in the <SearchSuggestField />
constructor to get it to initialize its state.