♻️(courses) refactor course and course run glimpse info into a state
Use the same state for the computation for the course and the
course run that can be called without the instance so that we can
use it also with the Elasticsearch results.

We added a distinction between a future course that is not yet open
for enrollment and a future course that is already closed for
enrollment. Indeed, the first one deserves a call to action.