♻️(search) use the page ID as Elasticsearch index
When categories move to a taxonomy, it will be much easier to get
ancestors IDs than getting the extensions' IDs. It also makes
more sense to use the page ID as it is the main DjangoCMS object
around which everything is built. So I refactored all objects to
use page IDs.