♻️(demosite) improve recursive page creation
The children of a page were set as an attribute of the parent page which
is a bit dirty. For the moment, let's just return all pages in a flat
dictionary and make sure they have a unique name. I also renamed the
argument `content` to `title` because that's what it is.