🚸(categories) limit categories that can be linked as plugin
When adding a category to a course via a plugin, we should only
propose leaf categories in the category tree. If the category
is a top category and has no children we can assume that it is
a meta category waiting for its childrenand should not propose it
even though it is technically a leaf.