♻️(search) refactor languages into a dynamic filter
We first implemented the languages filter as a hardcoded filter
with predefined choices in `defaults.py`. This was not perfect as
it prevented us from using all the possible language choices from
`settings.ALL_LANGUAGES`.

Instead, we can use a simple "terms" in ElasticSearch and treat
the field exactly like we already do categories and organizations.
The new `FILTERS_DYNAMIC` name also reflects the fact that it's not
just resources.