🐛(jobs) fix circle CI upload string with new file structure
As we refactored the app to move frontend related files into src/frontend,
we inadvertently broke the strings upload script run by the CI.
We did not notice it before as it is only ran on master, not on
feature branches.

We decided to improve the file structure there too, and properly manage
the dependencies in a package.json (as opposed to piggy-backing to
the frontend's dependencies).