♻️(dependencies) update types for new Jest typings
The Mock type in Jest changed to take as parameters the types of the
return value and the parameters of the mocked function instead of
taking directly the type of the mocked function.

We added our own helper `jestMockOf` to retain the earlier, less
verbose definition along with the new one.