✨(search) index related object names in course runs
We need to be able to surface courses with organization X or
category Y when a user does a full-text search for X or Y.
Currently, this does not happen if X and/or Y do not appear in
the syllabus or title of the course.

In order to enable this behavior, we first need to index these
informations on the course runs so we can search for them.