♻️(search) make the course query builder more generic
As we made the course query builder, we created a namedtuple,
"KeyFragmentPair", to help us constrain, manage & access our query
fragments throughout our code.

We are about to move some of this code into our new FilterDefinition
classes (as they are literally tasked with defining the filters,
this is part of their responsibility). Remove `KeyFragmentPair` so
we don't have to import it or pass it around to build query fragments.