♻️(search) build filter related queries/aggs in FilterDefinitions
FilterDefinitions implement a common interface to specify the way
a filter builds its filtering abilities as well as its own
aggregations in the context of a larger query.

This simplifies `CourseIndexer`'s `build_es_query` method and makes
it more generic and easier to change (as one simply has to change the
FilterDefinition methods which are way less intimidating!).

It also means we'll be able to implement the ES response consuming
logic right next to the ES query building for each kind of filter.