🚧(search) update defaults for FilterDefinitions
Filters from the defaults (or settings) will now operate from a
FilterDefinition subclass and its parameters, instead of flat
dictionaries with they way to consume them hardcoded in the
Course indexer and viewset.

Add a FILTERS/FILTERS_DEFAULT to `defaults.py`, which are the
replacement for FILTERS_DYNAMIC and FILTERS_HARDCODED. We have
to keep those around for now so we don't have to rework the "facets"
code in the Course list viewset response to use FilterDefinition,
as that whole part will be removed once the client is updated to
use "filters" instead.