♻️(frontend) rework frontend search use filters in courses response
Filter definitions are now directly available on the response object
for the course search API request. This means it is no longer necessary
to fetch disparate data and aggregate it into filter definitions in the
frontend.

Instead, we simply have to show whatever is returned by the server's
filter definitions and let users manipulate the query parameters
through our interface. Changes to the query parameters in turn trigger
new API requests, which yield both updated search results, and updated
information for the filters.

To achieve this, we realized we did not need the whole redux +
connected components machinery and could instead combine two custom hooks
with some light context use to have the frontend work as a simple yet
100% dynamic system.

This let us cut the bundle size ~30% and remove most of the complex code
from the frontend. We're now doing pure react with some raw JS helpers
where needed.