✨(search) manage search height during bootstrap & loading
Search shows a glitch during bootstrap/loading because the page
first loads and renders the React frontend without any content.

We can help this by setting a minimal height on the `body-content`
for the search page and growing the relevant elements to occupy
it. The filters column is then shown from load, we don't have the
footer jumping around, and we can show a loader on the course list
area.