📌(pylint) pin astroid dependency version
An update to `astroid` broke `pylint` in the CI and our own builds.
This can happen because `pylint` has a dependency on `astroid`, and
transitive dependencies are not pinned as there's no lockfile.

We can solve this (and resume our work & builds) by pinning a version
of `astroid` compatible with `pylint` in our own dependencies.