✨(search) respect the order of filter values as returned by the server
As we are about to add support for nested categories, it is impractical
to reorder filter values on the frontend side based on their activity
status: the underlying logic might be difficult to comprehend for users
and the reordering itself would make the UI jump around too much.

Instead, we can respect the order of the values in the API response. The
server can more easily order them based on the available information,
especially the published-defined order from the CMS itself.