🔥(search) remove FilterDefinitionWithValues type
As the filter definitions are henceforth generated on the backend side,
we do not need to juggle with two states for filter definitions, one
including computed values and the other one omitting them.

We can just keep `FilterDefinition` around and assume it includes the
values as it is the API's job to provide them to us.