✨(search) show nested filter values (lv. 1) in the search filters pane
Some filter values can have children filter values. In terms of search,
this means if the parent is selected all children should be matched, but
not the other way around.

We need to update our UI to be able to show this relation and easily
show/hide child filter values depending on the current state. We added
a Hook to encapsulate the common status behavior for parent & leaf
components.

We also had to rewrite the CSS to share most of the styles for the
relevant parts.