💄(project) add new template for a simple single column page
Currently there is only a "full width" page where content takes the
full width which is a little rude to make simple page like Annexes.

We need to add a new simple template which only include a single
column structure with a little vertical padding so we have a nicer
render for simple pages.