🔥(search) remove elasticsearch features tests
These tests are now covered by test_query_courses.py because we make
real Elasticsearch queries and test all our filtering and sorting.
It is difficult to Maintain these tests and make sure that we are
testing all the functionalities we use, and not the ones we don't use
anymore. So let's remove it.