✨(search) add a state field to the course document returned on the API
We already know how to compute this information on the course run
model. The difference here is that we need to compute it inside
Elasticsearch using the nested course run objects. We use a script
field and share the script source with the sorting script because
their code is mostly the same and there is no other way to factorize
scripts in Elasticsearch.