🐛(project) fix rebuild & bootstrap Makefile rules
When Richie is not running, the make migrate command does not up the
database (surprisingly), we should up the stack before running the
migrations. In fact, it makes sense to boot the stack in a bootstrap
command.

Also, we now rebuild separately the base and dev images, has we want to
pass a build argument to the dev build.