🐳(project) remove sources from base image
The base image is supposed to run Richie installed in the python
library, and thus does not require sources to be bundled in the
aforementioned image. With this constraint in mind, we removed the
back-end sources, and the front-end build is now bundled and installed
within the Richie package.

In the opposite, the development image requires an editable
installation, we thus removed globally installed the Richie application,
and re-installed it along with development dependencies.