🐛(search) update the frontend to the actual shape of records
Course dates were not displayed anymore as the date fields do not
exist on courses anymore and were replaced with a state field that
contains only the necessary data to display our date information
in the <CourseGlimpse />. We need to use that for our components to
render what we need.

Additionnally, IDs on records are strings, so we changed the type
and examples in tests to reflect this type.