🐛(project) fix watch-sass for sources out of frontend directory
Currently watch-sass does not see when a Sass source from an app is
changed and so it does not perform a new build. This would fix and
close #499

Fix only add `--watch` arguments to nodemon command to define every
directory to watch instead of default behavior to watch only current
directory.