🐛(courses) fix getting course from course run when below a parent page
The get_course method was broken when the course was below a root page,
because the query join was generating duplicate results. I added a
clause to fix it while still allowing the course to be at the root of
the page tree. Added some explanations and a specific test because it
is far from trivial.