🐛(demo) deactivate ES updates when creating the demo site
If we don't deactivate it, the ES index is being updated for each page
that is being created. We want the ES index to be built at the end,
once of the objects have been created in database. I introduced a
new setting to make it easy and avoid having to disconnect signals.