✨(search) split parts of the category taxonomy in several filters
We want to be able to define different filters for different parts
of the category taxonomy. This is achieved in 3 steps:
- allow pointing an indexable filter to a meta category,
- allow restricting facetting with a regex,
- apply the correct regex restriction for each indexable filter.
We now facet only on the top level of the taxonomy for each filter.
Facetting on children will have to be explicitly requested so the
the client can make specific requests when the user tries to drill
down on a category.