🐛(search) fix courses autocomplete
Autocomplete for courses was broken due to recent changes in the
way courses are formatted in ElasticSearch by our state script.

Courses from the autocomplete query do not use the state script and
are thefore missing expected fields when they are being formatted to
the shape we're returning on the API.

We reasoned this is not unwanted behavior: we don't need the same
shape for an autocomplete item as a full item, as we're only making
use of IDs, titles and, in the case of courses, absolute_url.

We can add a `format_es_document_for_autocomplete` on indexers that
builds these lighter objects for the sole use of autosuggest.