✨(search) allow limiting what facets are computed by indexable filters
Facets on related page objects (e.g. categories or organizations) are
now computed only for the top level objects below the meta page. In order
to display the children with their facet counts, the client needs to make
a separate query. This is what this commit allows: limiting the facets
that are computed by passing a regex in the querystring for each indexable
filter.