🐛(search) improve caching of include regex on indexable filters
The include regex was cached on the first execution, whatever the
result. In cases where the setting was set before the page was
created and published, it could cause the method to cache the
empty regex. In this case, we had to restart the app (on all pods
in case of scaled deployments) for the include regex to be
updated and for the filters to show correctly.

The solution applied here is to cache the result only when the
page pointed by the setting is found. Until then, the cache is
not set and we keep looking for the page.