🐛(search) make sure selected filter values are included in the result
On indexable filters that are using a terms aggregation with a limit
of 10 facet results, we could end up with a select filter value not
being included in the top 10 facets. Since the frontend now relies
on the query result to display the filter boxes, the selected facet
was not visible on the site anymore and could not be unselected...

We fix this problem by forcing specific facetting on the selected
filter values. We choose to merge them with the facets returned by
the terms aggregation. It can lead to more than 10 facets but we
think this allows keeping a significant number of new available
choices.

Fixes #504