✨(social-badges) add social network badges template tag
We need a way to configure common social network badges and to
include them in any template.

This add a new setting "SOCIAL_NETWORKS_BADGES" to define badge
configurations than can be called from a template tag
"{% badges ... %}" and rendered. Badge context is able to manage
text translation and stand on Django template syntax to be flexible
enough.