✨(search) improve fulltext search to match partial words
We use N-grams but only on the index, not the query because it results
in nearly all the courses to match any query (Google style) and the
sorting is how the most important results are returned first. But this
does not work for us because we force sorting within "bucket" depending
on the state of course runs.

In our case it is better to exclude results that don't match the query
close enough. A first approach is to consider the full query and not
break it down to n-grams. We will improve later based on practical
feedback.

Co-authored-by: Mehdi Benadda <me@mbenadda.com>