✨(search) implement parent & children filters in the frontend
The existing implementation of parent & children filters in the
frontend expected filter values to have a "children" value that
directly contained their own children values.

We decided to go another way with standard formatted paths that
give us many benefits in terms of search but needed a different
implementation in the frontend as parent filters need to fetch their
children when the user requests them. This also improves scalability.