💄(search) fix search filters glitch on click
When clicking on a search filter to toggle it, search filters lists are
updated producing a glitch in the search filters column. We've decided
to make each search filter group list a block with a fixed height to
avoid this disturbing glitch (blocks jump).

An alternative would have been to define a max-height, but this would
just limit the jump phenomenon that will still be disturbing when only a
few filters are still available.