🐛(organizations) add empty image fillers on organization detail
Currently organization detail template rely on a image that does not
exists when logo placeholder is empty, which leads to some errors on
CI and display.

This remove usage of this unexisting image, instead just use a empty
div with basic styles to display something empty. This has been done
for logo and banner placeholders.