I18n support (#566)
* ✨(i18n) configure i18n to manage translations

We use crowdin to manage translations. In order to automate the all
process we need to configure the crowdin-cli and add some helpers in ou
Makefile.

* ♻️(CI) refactor CI to upload source translations

The CI needs to be refactor to upload source translations when a branch
is merged on master.

* 🔥(tooling) remove scripts folder

Scripts folder contained our handcrafted tool to upload files to
crowdin. We now use the official CLI and we don't need anymore this
tool.

* 🌐(i18n) add translation files

We want to commit the translated messages in the project. Here are the
first files coming from crowdin.

* ✨(front) load translations with react-intl

React-intl was installed but no translations were used. Now, based on
the language code property we are able to load translations or fallback on
default messages if translation is missing.

* ✨(i18n) add setting and template filter to configure React locales

React only works with ISO15897 locales (e.g. fr_FR). Django works with
simple languages (e.g. fr) or full languages (e.g. fr-fr). We need
to make sure that React is bootstraped with the good locale whatever
the way we chose to define languages in Django.