📦(dependencies) refactor app/sandbox dependencies
Richie is supposed to be distributed as a django third-party
application, we needed to separate direct dependencies from runtime
dependencies (sandbox).

Also, pinning direct dependencies for the app is not a good idea since
we do not need to force a particular release when it's not required.
This will ease the app installation by avoiding unresolvable
dependencies.

Finally secondary dependencies should not be listed nor pinned as their
constraints should be declared in primary dependencies' package.