🚸(courses) allow snapshotting course when creating new course runs
Snapshotting courses is hidden in the menu and could be overlooked.
This may lead to courses not being backed-up when they should and
history getting lost. By adding a flag in the form that creates new
course runs, we make sure that users are always reminded of the
possibility to back-up their course right at the moment they are
the most likely to start modifying the current version.