✨(search) add base path to filter information
The autocomplete retrieves suggestions and needs a way to assign the
results to one of the filter boxes in order to select the filter
value when the suggestion is clicked. A clean and easy way to do
this is to add the path of the base page to the filter box information.