♻️(search) fix category selection in autocomplete
Selecting a category in the autocomplete suggestions caused the app
to crash. This was due to a mismatch between the filters object, that
contained definitions for meta-categories, and the autocomplete code, which
was looking for a "category" definition.

This can be fixed by adding the path to filter definitions and searching
in the filters object for a definition that matches the beginning of the
filter we're attempting to add.