✨(courses) add a new "assessment" section in the course page
We were missing a section to describe how students are being evaluated
on the course. We don't display anything on the published page if this
section is left empty. We had to introduce a new template filter to
do this because the `page_placeholder` template tag still returns
some markup even when it has no content (linked to placeholder handling
via js).