♻️(frontend) improve the react-intl mocks
We first used mocks of react-intl components when testing with enzyme
because it was highly impractical to wrap every component in the
<IntlProvider />. It messed up shallow rendering and was adding
boilerplate to get to the actual components when it even worked.

Now that we gave up shallow rendering and use react-testing-library, we
don't have the same need for the mock. It's still a bit of boilerplate to
wrap any intl-using components in IntlProvider in tests but it will
allow us to be closer to real conditions in our tests.