✨(search) add a helper do remove all filters at once
When the user selects many filters, it can be painful to go back and
remove them one by one. Especially since they are persisted in the
URL: reloading the page does not clear them either.

We can simply add a link in the filters pane so users can remove
everything at once and make another search.