♻️(search) remove redundant unit tests
Resulting filter values from FILTER_ADD & FILTER_REMOVE operations
were tested both in a "pure" test for `computeNewFilterValue` and in
a more functional test for `useCourseSearchParams`. This is painful
when we want to cover more cases as we have to do it 2 times to be
consistent.

In our opinion the functional tests are way more valuable as they're
closer to real-world use, and in this instance also test the resulting
query string. We added to the hook tests some cases that were only
covered in `computeNewFilterValue` tests. We can then get rid of those
`computeNewFilterValue` tests.