✨(search) remove an active filter when adding its parent or child
When a parent filter is selected and we select one of its children,
it is confusing that nothing happens and that courses shown are still
for ALL the parent's children.

Conversely, when a child filter is selected and we select the parent,
it is confusing that the child filter is now useless.

In those cases, we can prioritize the new value's expected outcome and
remove the existing value that conflicts with that outcome: this means
removing the existing child when its parent is activated, and removing
the existing parent when its child is activated.