♻️(demo) move demo to its own app
The command that creates the demo site was inside the core app. This
was creating circular dependencies issues and was bad design because
the dependency tree should be in order: core <- courses <- demo.

I took the chance of this demo to remove real images from tests and
rely only on factory boy's dummy images. Real images are relocated
to the demo app and only used there.