🔍(SEO) configure meta and opengraph tags
There is no currently any meta tags in page, this is an issue for
SEO.

Here we add and define meta tags as meaningfully as possible for
home and detail pages. To do so we add new dedicated blocks in base
template with some basic informations then page templates can fill
or override them to define relevant meta informations.