✨(search) add the name of the organization to show on api responses
Course glimpses in search results need to show the name of a related
organization: either the main one (1st in order in the list of linked
organizations on the CMS), or the one that was deemed most interesting
for the current search.

To do this, we first need to add this relevant organization's name
to the API responses for course searches.