✨(search) show the highlighted organization on the course glimpse
Course glimpses throughout Richie show the highlighted organization
for the course (which, for now, is always the organization at index 0
in the list of organizations linked to the course).

This was missing in the search results because the data was not
included in the API. Now that it is, we can easily show it in the
course glimpses.