✨(search) course glimpse footer in search matches CMS version
As we continue to update the course glimpse in Search to match
the ones the CMS is producing, we need to add CTAs and manage
footer text/date like it is done in the CMS side (via Django
templates).

We needed to remove the `display: flex` from the date element as
`react-intl` creates a `<span />` when it generates the human
representation of the date, which was being messed up by the
flex.