💄(search) show a placeholder when a course is missing a cover
When a course is missing a cover image, course glimpses in the search
results were shown with a broken aspect ratio because they were
simply showing an empty image tag.

We're simply replacing that with a "Cover" placeholder on a grey
background so instructors see what's missing and can fix it quickly.
It also makes unfinished pages look better while instructors are
still working on them.