♻️(search) refactor SearchSuggestField to use more functional tests
<SearchSuggestField /> was using tests that were very "unit" oriented,
testing disjointed parts with a lot of mocks to glue everything. This
is not something we favor anymore as it does not make it simple to
cover use-cases and actual reported bugs with new tests.

We decided to refactor the tests to work in a more functional way by
mimicking user interaction with `react-testing-library`. Our "unit"
now being <SearchSuggestField /> instead of the various functions it
has and APIs of `react-autosuggest` it intereacts with.

As we did this, we noticed functions that did not need to be exported
and curried anymore, and we did some light changes that simplified the
actual component code too.