🐛(search) fix a display glitch in active parent filters
Active parent filters in the search filters pane were displayed with
2 borders due to a mistake in the styles (as most styles are common
between parent self filters & leaf filters).

We updated the CSS to fix this glitch and reorganized & added comments
to the `.scss` file to make it more explicit where the two uses for
these styles need to be handled with special care.