🐛(circleci) fix the hub job
The build-docker* jobs are using CircleCI docker layers caching feature.
At first, I thought that the images will also get restored in future
jobs, but this is not the case; only layers are cached. Thus, in order
to properly tag and publish build images, we need to rebuild them first
(using cached layers). I expect that no new layer will be created and it
should be blazing fast.