✨(search) use checkboxes to enable/disable filters in search pane
We used button with a close (X) icon to enable users to manage active
filters in the filters pane of the search view. This was not the
best way to help users intuitively understand what those checkboxes
enable them to achieve.

Using checkbox, they can understand without any explanation that ticked
filters are included in the results and non-ticked filters are not.
This is also very helpful to make the way multi-faceted filters (OR)
work obvious.