✨(search) disable course suggestions in the SearchSuggestField
We're implementing "live search", that is, full-text search that
happens as the user is typing. This should give users very quick
feedback with relation to what they're looking for in courses, and
make autosuggest unnecessary for this use-case.

We're removing it for the <SearchSuggestField /> in course search
but might end up reusing the same logic in another component used
to a more global search in Richie, therefore we removed it from
the component but not from the helpers in `utils`.