🏷️(autosuggest) improve typings in SearchSuggestField
The functions passed as props to <Autosuggest /> were manually typed
at their definition site. This is fine from a correctness perspective
as they're typechecked against the expected props when we pass them.

It is however redundant, as <Autosuggest /> already provides typings
for its props that we can reuse. This also makes it easier to introspect
eg. arguments to callbacks as those typings are complete, as opposed to
the minimal ones we define with our own functions.