✨(frontend) search for courses as the user types
Instead of offering autocomplete-based suggestions, we can offer users
a more intuitive and faster experience by directly querying the
search index itself with their query as they are typing.

This is what we're enabling here. We can simply rely on our existing
change handler for the input, the only differences being checking for
the change method to only run a search when it comes from the user (not
when they eg. hover a suggestion), and debouncing these requests to
avoid overloading the server.